Changed a large portion of the underlying logic to align with monster "classes" (bomber, nuker, healer, etc) to better match which sets / which stats on those sets are most important to the archetype.

+++ src/org/DemonSkye/wut/Main.java
@@ -26,7 +26,7 @@ public class Main {
             FileOut.write("scroll");
         }
         if (str.contains("Rune")){
-            Integer runeReturn = RuneDrop.runeEval(str);
+            Double runeReturn = RuneDrop.runeEval(str);
             if(runeReturn == 2){
                 System.out.println("runeReturn get: " + runeReturn);
                 found = true;
